+++ nextgisweb/resource/nodepkg/component/resource-picker/ResourcePickerFooter.tsx
@@ -1,7 +1,7 @@
 import { observer } from "mobx-react-lite";
 import { useEffect, useMemo, useRef, useState } from "react";
 
-import { Button, Col, Input, Row, Space, Tooltip } from "@nextgisweb/gui/antd";
+import { Badge, Button, Col, Input, Row, Space } from "@nextgisweb/gui/antd";
 import type { InputRef } from "@nextgisweb/gui/antd";
 import { errorModal } from "@nextgisweb/gui/error";
 import type { ApiError } from "@nextgisweb/gui/error/type";
@@ -138,6 +138,7 @@ function MoveControlInner<V extends SelectValue = SelectValue>({
     };
 
     const OkBtn = ({ disabled }: { disabled?: boolean }) => {
+        const badgeCnt = multiple && selected.length > 1 ? selected.length : 0;
         return (
             <Button
                 type="primary"
@@ -149,6 +150,9 @@ function MoveControlInner<V extends SelectValue = SelectValue>({
                 }}
             >
                 {getSelectedMsg}
+                {!!badgeCnt && (
+                    <Badge size="small" color="transparent" count={badgeCnt} />
+                )}
             </Button>
         );
     };
@@ -159,27 +163,25 @@ function MoveControlInner<V extends SelectValue = SelectValue>({
                 {allowCreateResource &&
                     possibleToCreate &&
                     !loading.createNewGroup && (
-                        <Tooltip title={msgCreateGroup}>
-                            <a
-                                style={{ fontSize: "1.5rem" }}
-                                onClick={onCreateClick}
-                            >
-                                <CreateNewFolder />
-                            </a>
-                        </Tooltip>
+                        <a
+                            style={{ fontSize: "1.5rem" }}
+                            onClick={onCreateClick}
+                            title={msgCreateGroup}
+                        >
+                            <CreateNewFolder />
+                        </a>
                     )}
             </Col>
             <Col>
                 {selected.length ? (
                     <Space>
-                        <Tooltip title={msgClearSelection}>
-                            <Button
-                                icon={<HighlightOff />}
-                                onClick={() => {
-                                    store.clearSelection();
-                                }}
-                            ></Button>
-                        </Tooltip>
+                        <Button
+                            icon={<HighlightOff />}
+                            title={msgClearSelection}
+                            onClick={() => {
+                                store.clearSelection();
+                            }}
+                        ></Button>
                         <OkBtn />
                     </Space>
                 ) : pickThisGroupAllowed ? (